feat(Movement): Movement 인터페이스 생성
